requirements.txt
beautifulsoup4==4.9.3
requests==2.25.1
mysql-connector-python==8.0.32